Phase 3 M8: PNG decode via the png crate behind a ported stb walk
Acceptance stays stb's, not the crate's: quake-image::png_stb hand-ports
stbi__parse_png_file's chunk walk (IHDR/PLTE/tRNS/IDAT/IEND rules, the
1<<24 and 1<<30 guards, unknown-critical rejects incl. the 4-char reason,
the post-read zeof quirk that rejects any <=2-byte zlib stream) plus
stbi__parse_zlib_header, then reassembles a canonical stream (zlib header
normalized to cinfo=7 -- stb ignores the declared window, fdeflate does
not) and uses png 0.18 for inflate/defilter/expand only, checksums off
(stb checks neither CRC nor Adler). Gray scaling and tRNS colorkeys are
proven collision-free vs stb's depth-scale table. CgBI PNGs route to the
C fallback wholesale. 16-bit reduces by high byte like stbi__convert_16_to_8.

First third-party runtime deps in libquake_rs.a: png + bitflags/crc32fast/
fdeflate/flate2/miniz_oxide/simd-adler32/adler2/cfg-if, all MIT-dual
(cargo deny green, ADR-003 amendment lands with the docs commit).

image_crate_differential grows the PNG matrix (13 color/depth combos x
interlace, tRNS variants incl. sub-byte and 16-bit keys, checksum/trailing
tolerance, split+empty IDAT, cinfo=15 acceptance, 20 structural reject
reasons compared exactly, crate-side rejects compared with the
owner-approved masked-reason policy); four mutations (LSB-for-MSB, broken
gray replication, normalization dropped, 2-byte zlib accepted) each fail
the suite. The real vq_pak PNGs now decode through the Rust path and
run_corpus.py --check passes 10/10 on build-rs (every launch decodes the
embedded PNGs). fuzz_image_stb added (60 s soak clean).

+++ rust/quake-image/src/tga.rs
@@ -363,7 +363,7 @@ mod tests {
         // R=31, G=0, B=15 -> 0b0111110000001111
         f.extend_from_slice(&0b0111_1100_0000_1111u16.to_le_bytes());
         let t = decode(&f).unwrap();
-        assert_eq!(&t.rgba, &[255, 0, (15 * 255) / 31, 255]);
+        assert_eq!(&t.rgba, &[255, 0, ((15u32 * 255) / 31) as u8, 255]);
     }
 
     #[test]
